+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Changed
 * **Breaking**: all setup hooks have been removed from the `packages` flake
   output. They can still be accessed via the `lib` flake output.
+* **Breaking**: `cargoBuild` now only runs `cargo build` in a workspace, tests
+  are no longer run
 * **Breaking**: `buildDepsOnly` does not automatically imply the `--all-targets`
   flag when invoking `cargo check`. Use `cargoCheckExtraArgs` to control this
 * `buildDepsOnly` now accepts `cargoCheckExtraArgs` for passing additional