+++ CHANGELOG.md
@@ -6,6 +6,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## Unreleased
 
+## [0.9.0] - 2022-10-29
+
 ### Changed
 * **Breaking**: all setup hooks have been removed from the `packages` flake
   output. They can still be accessed via the `lib` flake output.
@@ -260,6 +262,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## 0.1.0 - 2022-01-22
 - First release
 
+[0.9.0]: https://github.com/ipetkov/crane/compare/v0.8.0...v0.9.0
 [0.8.0]: https://github.com/ipetkov/crane/compare/v0.7.0...v0.8.0
 [0.7.0]: https://github.com/ipetkov/crane/compare/v0.6.0...v0.7.0
 [0.6.0]: https://github.com/ipetkov/crane/compare/v0.5.1...v0.6.0